requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
dacite==1.8.1
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pokemontcgsdk==3.4.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
